pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipython>=9.3.0",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
